fix gcc 4.3 warning
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@60245 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 6b10a6376f5625ba11af34def0e265e69f0e97bb

+++ LHCb/L0/ProcessorKernel/src/TileRegister.cpp
@@ -136,7 +136,7 @@ void L0Muon::TileRegister::print_bits(unsigned int event, FILE *file){
       
       if (mid.isValid()){
         fprintf(file,"%5u %3u ( %2d,%2d ) Q%1d M%1d R%1d %2d %2d  %d\n"
-                ,event ,(unsigned long)i 
+                ,event ,(unsigned int)i 
                 ,mid.layout().xGrid()
                 ,mid.layout().yGrid()
                 ,mid.quarter()+1 
@@ -147,7 +147,7 @@ void L0Muon::TileRegister::print_bits(unsigned int event, FILE *file){
                 ,val
                 ); 
       } else {
-        fprintf(file,"%5u %3u                           %d\n",event,i,val);
+        fprintf(file,"%5u %3u                           %d\n",event,(unsigned int)i,val);
       }
     }
